perf(bundle): lazy-load persistence wiring + payload onto the async path
Persistence's wiring closure (wire-persistence.ts) and the heavy payload
machinery (the onFormChange-driven debounced writer, envelope read/build,
pluck / strip / filter) were statically imported into use-abstract-form,
so they shipped on the always-on useForm path even though persist: is a
strict opt-in. Move them behind a dynamic import so a minimal useForm
consumer never pays for them.

What stays eager (core-anchored): the orphan-sweep machinery +
getStorageAdapter + config helpers (every client form runs the sweep to
clean stale keys after a dropped persist:), sensitive-names and
opt-in-registry (built in create-form-store, read on the render path), and
mergeSparseHydration / mergeDeep (create-form-store's SSR rehydrate imports
them). The movable wire-persistence-exclusive code is extracted into a new
lazy persistence/payload.ts that only wire-persistence imports.

No new async for any synchronous caller (the binding constraint):

- Render path unchanged. register-api throws on register({ persist: true })
  for a form without persist: via state.modules.has(PERSISTENCE_MODULE_KEY).
  The key now holds a PersistenceHandle { config, ready } set SYNCHRONOUSLY
  at mount, so .has() and the divergence warn (reads handle.config) still
  answer synchronously. wiredConfig is dropped from PersistenceModule (its
  one reader, the divergence warn, moved to the eager handle).

- Flash-of-defaults window unchanged. The caller starts the adapter import
  (getStorageAdapter, still eager) in parallel with the wire-persistence
  chunk import and injects the adapter promise, so the chunk load hides
  behind the adapter load the hydration read already awaited. wirePersistence
  takes the adapter promise as a required arg.

- form.persist / clearPersistedDraft were already async; they now await
  handle.ready (resolves to the module, or undefined on chunk-load failure
  / disposal). reset / resetField stay synchronous-returning with the wipe
  still fire-and-forget at the same wall-clock.

Disposal race guarded like D1: an eager registerCleanup flips a
formDisposed flag the resolver checks before wiring onto a dead store; a
failed chunk load degrades to persistence-unavailable, never throwing into
the consumer.

Tests route through useForm, so the persist-write subscription now attaches
a few microtasks after mount (once the chunk lands) rather than
synchronously. A new waitForPersistence(app) harness helper awaits every
form's handle.ready; suites that dispatch input immediately after mount
(modelling a keystroke faster than a real user can produce) establish
first, and two fake-timer suites pre-warm the chunk so the wiring IIFE
resolves on a microtask.

Eager set drops 45.61 -> 44.60 kB gz (check-eager-size budget tightened to
lock it). The size-limit total caps measure the inlined bundle (no
splitting, so the dynamic import folds back in), so zod-v4 ticks 32 B over
its cap; bumped 56 -> 57 KB with the reason recorded. Full suite
305 files / 3698 pass.
