+++ scripts/check-eager-size.mjs
@@ -154,12 +154,15 @@ export async function measureEager(define = PROD_DEFINE) {
 // Committed eager budget (gz bytes) for a minimal `useForm` (zod-v4).
 // Baseline measured at 46.28 kB gz when this gate landed, with the
 // dev-flag DCE win (core/dev.ts) folded in under the production define.
-// D1 then lazy-loads multi-tab sync onto the async path, dropping the
-// eager set to 45.61 kB gz; the budget is tightened here to lock that
-// in. ~0.5 kB headroom absorbs minifier-version drift. The lazy-loading
-// work tightens this as optional features move to the async path; never
-// loosen it without a recorded reason in the commit.
-const BUDGET_GZ = 47_200
+// D1 then lazy-loads multi-tab sync onto the async path (45.61 kB gz),
+// and D2 lazy-loads persistence's wiring + payload machinery (the
+// onFormChange writer, envelope read/build, debounce, pluck / strip /
+// filter) onto the async path, dropping the eager set to 44.60 kB gz;
+// the budget is tightened here to lock that in. ~0.5 kB headroom absorbs
+// minifier-version drift. The lazy-loading work tightens this as
+// optional features move to the async path; never loosen it without a
+// recorded reason in the commit.
+const BUDGET_GZ = 46_200
 
 const isMain = import.meta.url === pathToFileURL(realpathSync(argv[1])).href
 if (isMain) {
